chore: rename utilities folder to utils to match convention
