+++ packages/extension-table/src/utils/parseAlign.ts
